add new applied field to config store
